refactor(ui): rename plot components
