+++ ui/src/shared/components/ViewSwitcher.tsx
@@ -6,11 +6,11 @@ import {Plot, FromFluxResult} from '@influxdata/giraffe'
 import GaugeChart from 'src/shared/components/GaugeChart'
 import SingleStat from 'src/shared/components/SingleStat'
 import TableGraphs from 'src/shared/components/tables/TableGraphs'
-import HistogramContainer from 'src/shared/components/HistogramContainer'
-import HeatmapContainer from 'src/shared/components/HeatmapContainer'
+import HistogramPlot from 'src/shared/components/HistogramPlot'
+import HeatmapPlot from 'src/shared/components/HeatmapPlot'
 import FluxTablesTransform from 'src/shared/components/FluxTablesTransform'
-import XYContainer from 'src/shared/components/XYContainer'
-import ScatterContainer from 'src/shared/components/ScatterContainer'
+import XYPlot from 'src/shared/components/XYPlot'
+import ScatterPlot from 'src/shared/components/ScatterPlot'
 import LatestValueTransform from 'src/shared/components/LatestValueTransform'
 
 // Types
@@ -73,7 +73,7 @@ const ViewSwitcher: FunctionComponent<Props> = ({
 
     case ViewType.XY:
       return (
-        <XYContainer
+        <XYPlot
           table={table}
           fluxGroupKeyUnion={fluxGroupKeyUnion}
           viewProperties={properties}
@@ -81,7 +81,7 @@ const ViewSwitcher: FunctionComponent<Props> = ({
           timeZone={timeZone}
         >
           {config => <Plot config={config} />}
-        </XYContainer>
+        </XYPlot>
       )
 
     case ViewType.LinePlusSingleStat:
@@ -99,7 +99,7 @@ const ViewSwitcher: FunctionComponent<Props> = ({
       } as SingleStatView
 
       return (
-        <XYContainer
+        <XYPlot
           table={table}
           fluxGroupKeyUnion={fluxGroupKeyUnion}
           viewProperties={xyProperties}
@@ -118,43 +118,43 @@ const ViewSwitcher: FunctionComponent<Props> = ({
               </LatestValueTransform>
             </Plot>
           )}
-        </XYContainer>
+        </XYPlot>
       )
 
     case ViewType.Histogram:
       return (
-        <HistogramContainer
+        <HistogramPlot
           table={table}
           loading={loading}
           timeZone={timeZone}
           viewProperties={properties}
         >
           {config => <Plot config={config} />}
-        </HistogramContainer>
+        </HistogramPlot>
       )
 
     case ViewType.Heatmap:
       return (
-        <HeatmapContainer
+        <HeatmapPlot
           table={table}
           loading={loading}
           timeZone={timeZone}
           viewProperties={properties}
         >
           {config => <Plot config={config} />}
-        </HeatmapContainer>
+        </HeatmapPlot>
       )
 
     case ViewType.Scatter:
       return (
-        <ScatterContainer
+        <ScatterPlot
           table={table}
           loading={loading}
           viewProperties={properties}
           timeZone={timeZone}
         >
           {config => <Plot config={config} />}
-        </ScatterContainer>
+        </ScatterPlot>
       )
 
     default:
